# bff/openapi.yaml
...
openapi: 3.1.0
info:
  title: Miaoyomi API
  version: 0.19.0
  description: "Everything the web app does, it does over this API. Session JWTs are for the browser; scripts
    use a personal API token (Profile -> Security -> API tokens), sent as `Authorization: Bearer uy_...`.
    Scopes: read (every token), write (anything that changes data), admin (the /api/admin surface, which also
    needs an admin account). /img/* takes the yomi_img cookie or the OPDS token over HTTP Basic; /opds/* takes
    the OPDS token over HTTP Basic; neither accepts API tokens. List endpoints answer { content: [...] };
    errors answer a non-2xx status with { error, message }. The prose guide with worked examples is
    docs/api.md."
servers:
  - url: /
    description: This server
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: A session JWT from POST /auth/login, or a personal API token (uy_...). Admin routes need the
        admin scope and an admin account.
    opdsBasic:
      type: http
      scheme: basic
      description: HTTP Basic where the password is the OPDS token from POST /api/opds/token.
    imgCookie:
      type: apiKey
      in: cookie
      name: yomi_img
      description: Set at login so <img> tags can authenticate.
...
paths:
...
  /api/setup:
    post:
      tags:
        - auth
      summary: Create the first admin account (only while no account exists).
      operationId: postApiSetup
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                displayName:
                  type: string
      security: []
      responses:
        "200":
          description: OK
        "400":
          description: Malformed body or parameter
        "409":
          description: "Conflict: already exists, already running, or duplicate"
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Log in with username and password (and a TOTP code when enabled); returns a short-lived JWT and sets
        the refresh cookie.
      operationId: postAuthLogin
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                totp:
                  type: string
      security: []
      responses:
        "200":
          description: OK
        "400":
          description: Malformed body or parameter
        "401":
          description: Not authenticated
        "429":
          description: Too many attempts
  /auth/register:
    post:
      tags:
        - auth
      summary: Create an account, when registration is open.
      operationId: postAuthRegister
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                displayName:
                  type: string
      security: []
      responses:
        "200":
          description: OK
        "400":
          description: Malformed body or parameter
        "403":
          description: "Forbidden: token scope, account permission or age limit"
        "409":
          description: "Conflict: already exists, already running, or duplicate"
  /auth/refresh:
    post:
      tags:
        - auth
      summary: Exchange the rotating refresh cookie for a new JWT.
      operationId: postAuthRefresh
      security: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /auth/logout:
    post:
      tags:
        - auth
      summary: End this session and clear its cookies.
      operationId: postAuthLogout
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /auth/logout-all:
    post:
      tags:
        - auth
      summary: Revoke every session of the calling account.
      operationId: postAuthLogoutAll
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /auth/me:
    get:
      tags:
        - auth
      summary: "The calling account: id, username, role, permissions."
      operationId: getAuthMe
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /auth/password:
    post:
      tags:
        - auth
      summary: Change the calling account's password.
      operationId: postAuthPassword
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                current:
                  type: string
                next:
                  type: string
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "400":
          description: Malformed body or parameter
        "401":
          description: Not authenticated
  /auth/sessions:
    get:
      tags:
        - auth
      summary: Active sessions of the calling account.
      operationId: getAuthSessions
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /auth/sessions/{id}:
    delete:
      tags:
        - auth
      summary: Revoke one session.
      operationId: deleteAuthSessionsId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
        "404":
          description: Not found, or not visible to this account
  /auth/totp/setup:
    post:
      tags:
        - auth
      summary: "Start two-factor enrolment: returns the secret and QR payload."
      operationId: postAuthTotpSetup
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "401":
          description: Not authenticated
  /auth/totp/enable:
    post:
      tags:
        - auth
      summary: Confirm two-factor with a first code; returns recovery codes once.
      operationId: postAuthTotpEnable
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "400":
          description: Malformed body or parameter
        "401":
          description: Not authenticated
  /auth/totp/disable:
    post:
      tags:
        - auth
      summary: Turn two-factor off (requires a current code).
      operationId: postAuthTotpDisable
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      security:
        - bearer: []
      responses:
        "200":
          description: OK
        "400":
          description: Malformed body or parameter
        "401":
          description: Not authenticated
...